requirements.txt
docker==5.0.3
fastapi==0.75.0
jsonpickle==2.1.0
lark==1.1.2
pytest==7.1.0
PyYAML==6.0
redisgraph==2.4.4
rq==1.10.1
uvicorn==0.17.6
